requirements/production.txt
blessings==1.6.1
bpython==0.17.1
certifi==2018.1.18
chardet==3.0.4
curtsies==0.3.0
dj-database-url==0.5.0
Django==2.0.2
django-cors-headers==2.2.0
django-model-utils==3.1.1
djangorestframework==3.7.7
drf-nested-routers==0.90.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
Pygments==2.2.0
python-decouple==3.1
pytz==2018.3
requests==2.18.4
six==1.11.0
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.1
